+++ src/components/Projects.tsx
@@ -1,25 +1,18 @@
 import clsx from "clsx";
 import { Tag } from "~/components/ui/Tag";
 import { type Filter } from "~/hooks/useFilter";
-import { impactCategoryLabels, type Project } from "~/hooks/useProjects";
-import { createComponent } from "./ui";
-import { tv } from "tailwind-variants";
-import { Card } from "~/components/ui/Card";
-import { CardTitle } from "./ui/CardTitle";
+import { type Project } from "~/hooks/useProjects";
+import { Card, CardTitle } from "~/components/ui/Card";
 import { ProjectImage } from "./ui/ProjectImage";
 import { ImpactCategories } from "./ui/ImpactCategories";
+import { Grid } from "./Grid";
 
 type Props = { filter?: Filter; projects?: Project[] };
 
 export const Projects = ({ filter, projects }: Props) => {
   const isList = filter?.display === "list";
   return (
-    <div
-      className={clsx("mb-8 grid gap-4", {
-        ["md:grid-cols-3"]: !isList,
-        ["gap-6 divide-y divide-neutral-200"]: isList,
-      })}
-    >
+    <Grid isList={isList}>
       {projects?.map((project) =>
         isList ? (
           <ProjectListItem key={project.id} project={project} />
@@ -27,7 +20,7 @@ export const Projects = ({ filter, projects }: Props) => {
           <ProjectGridItem key={project.id} project={project} />
         )
       )}
-    </div>
+    </Grid>
   );
 };
 
@@ -41,7 +34,7 @@ export const ProjectGridItem = ({ project }: { project: Project }) => {
         </div>
         <CardTitle>{project.displayName}</CardTitle>
         <p className="line-clamp-2 text-sm text-neutral-700">{project.bio}</p>
-        <ImpactCategories tags={project.impactCategory} />
+        <ImpactCategories tags={project.impactCategory} type={"projects"} />
       </div>
     </Card>
   );
@@ -56,10 +49,8 @@ export const ProjectListItem = ({ project }: { project: Project }) => {
         <p className="line-clamp-3 text-sm text-neutral-700 sm:line-clamp-2">
           {project.bio}
         </p>
-        <ImpactCategories tags={project.impactCategory} />
+        <ImpactCategories tags={project.impactCategory} type={"projects"} />
       </div>
     </div>
   );
 };
-
-
