--- src/components/ui/CardTitle.tsx
@@ -1,9 +1,0 @@
-import { tv } from "tailwind-variants";
-import { createComponent } from ".";
-
-export const CardTitle = createComponent(
-  "h3",
-  tv({
-    base: "text-lg font-bold",
-  })
-);
