+++ src/hooks/useProjects.ts
@@ -1,13 +1,7 @@
 import { useQuery } from "wagmi";
 import { initialFilter, type Filter } from "./useFilter";
-import { useMemo } from "react";
 import { projects } from "~/data/mock";
-
-export type ImpactCategory =
-  | "OP_STACK"
-  | "COLLECTIVE_GOVERNANCE"
-  | "DEVELOPER_ECOSYSTEM"
-  | "END_USER_EXPERIENCE_AND_ADOPTION";
+import { type ImpactCategory } from "./useCategories";
 
 export type Project = {
   id: string;
@@ -50,12 +44,6 @@ export type Project = {
   certifiedNotBarredFromParticipating: boolean;
 };
 
-export const impactCategoryLabels: { [key in ImpactCategory]: string } = {
-  COLLECTIVE_GOVERNANCE: "Collective Governance",
-  OP_STACK: "OP Stack",
-  DEVELOPER_ECOSYSTEM: "Developer Ecosystem",
-  END_USER_EXPERIENCE_AND_ADOPTION: "End user UX",
-};
 
 export function useProjects(filter: Filter) {
   const {
@@ -94,16 +82,4 @@ export function useProjects(filter: Filter) {
   );
 }
 
-export function useCategories() {
-  return useMemo(() => {
-    // Set each category to 0 - { OP_STACK: 0, COLLECTIVE_GOVERNANCE: 0, ...}
-    const initialState = Object.keys(impactCategoryLabels).reduce(
-      (a, x) => ({ ...a, [x]: 0 }),
-      {}
-    );
-    return projects.reduce((acc, x) => {
-      x.impactCategory.forEach((category) => (acc[category] += 1));
-      return acc;
-    }, initialState as { [key in ImpactCategory]: number });
-  }, [projects]);
-}
+
